// src/Api/models/blog.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Hey Pharamcist API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface Blog
 */
export interface Blog {
  _id?: string;
    /**
     * 
     * @type {Date}
     * @memberof Blog
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof Blog
     */
    updatedAt: Date;
    /**
     * 
     * @type {string}
     * @memberof Blog
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Blog
     */
    blogTitle: string;
    /**
     * 
     * @type {string}
     * @memberof Blog
     */
    image: string;
    /**
     * 
     * @type {number}
     * @memberof Blog
     */
    order: number;
    /**
     * 
     * @type {string}
     * @memberof Blog
     */
    content: string;
    /**
     * 
     * @type {boolean}
     * @memberof Blog
     */
    isActive: boolean;
    /**
     * 
     * @type {string}
     * @memberof Blog
     */
    storeId: string;
}