// src/Api/models/update-items-order-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Hey Pharamcist API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface UpdateItemsOrderDto
 */
export interface UpdateItemsOrderDto {
  _id?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof UpdateItemsOrderDto
     */
    ids: Array<string>;
}